consolidate normalizeTxParams in `tx`
